Admin user

+++ src/main/java/com/masai/service/UserServiceImpl.java
@@ -6,29 +6,64 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.masai.exceptions.AdminException;
 import com.masai.exceptions.UserException;
+import com.masai.model.CurrentAdminSession;
+import com.masai.model.CurrentUserSession;
 import com.masai.model.User;
+import com.masai.repository.AdminDao;
+import com.masai.repository.CurrentAdminSessionDao;
+import com.masai.repository.CurrentUserSessionDao;
 import com.masai.repository.UserDao;
 
 @Service
 public class UserServiceImpl  implements UserService {
 	@Autowired
 	private UserDao userDao ;
+	
+	@Autowired
+	private CurrentUserSessionDao usDao;
+	
+	@Autowired
+	private AdminDao aDao;
+	
+	@Autowired
+	private CurrentAdminSessionDao sDao;
 
 	@Override
-	public User addUser(User user) throws UserException  {
-		Optional<User> opt = userDao.findById(user.getUserLoginId()) ;
-		if(opt.isPresent()) {
-			throw new UserException("User already Present!") ;
-		}else {
-			return userDao.save(user) ;
+	public User addUser(User user,String key) throws UserException  {
+		CurrentAdminSession loggedInAdmin= sDao.findByUuid(key);
+		
+		if(loggedInAdmin == null) {
+			throw new AdminException("Please check key, No Admin loggedIn with given key");
+		}
+		else {
+			User opt= userDao.findByContact(user.getContact()) ;
+//			if(user.getUserLoginId() != loggedInUser.getUserId()) {
+//				throw new UserException("Invalid user details, Please Login first ");
+//			}
+			if(opt != null) {
+				throw new UserException("User already Present!") ;
+			}
+			else {
+				return userDao.save(user) ;
+			}
 		}
 	}
 
 	@Override
-	public User updateUser(User user) throws UserException {
+	public User updateUser(User user,String key) throws UserException {
+		CurrentUserSession loggedInUser= usDao.findByUuid(key);
+		
+		if(loggedInUser == null) {
+			throw new UserException("Please provide a valid key");
+		}
+		
 		Optional<User> opt=  userDao.findById(user.getUserLoginId()) ;
 		
+		if(opt.get().getUserLoginId() != loggedInUser.getUserId()) {
+			throw new UserException("Invalid user details, Please Login first ");
+		}
 		if(opt.isPresent()) {
 			return userDao.save(user) ;
 		}else {
@@ -38,8 +73,17 @@ public class UserServiceImpl  implements UserService {
 	}
 
 	@Override
-	public User deleteUser(int userId) throws UserException {
+	public User deleteUser(int userId,String key) throws UserException {
+		CurrentAdminSession loggedInAdmin= sDao.findByUuid(key);
+		
+		if(loggedInAdmin == null) {
+			throw new AdminException("Please check key, No Admin loggedIn with given key");
+		}
+		
 		Optional<User> opt=  userDao.findById(userId) ;
+//		if(opt.get().getUserLoginId() != loggedInUser.getUserId()) {
+//			throw new UserException("Invalid user details, Please Login first ");
+//		}
 		
 		if(opt.isPresent()) {
 			 userDao.deleteById(userId); ;
@@ -62,7 +106,7 @@ public class UserServiceImpl  implements UserService {
 	}
 
 	@Override
-	public List<User> viewallUsers() throws UserException {
+	public List<User> viewAllUsers() throws UserException {
 		List<User> allUsers =  userDao.findAll() ;
 		
 		if(allUsers.size() != 0) {
